Complete fix for OmitFromUsageDocs.
* PhotoAlbumManagerExpectedBrowserUsage.txt: Update expected text as empty
spaces differences are not significant for HTML rendering and doesn't seem
obviously related to my changes. It's yet unclear to me if these diffs need to
be investigated.

* UsageTests.cs: Add the case of a method that should be Omitted from usage doc.

* CommandLineAction.cs (UsageSummaryHTMLEncoded): loop in the UsageArguments and
therefore drop the test from within the loop.
(OmitFromUsage): New method indicating if the command should be omitted from
usage.

* CommandLineArgumentsDefinitions.cs (UsageActions): Returns only the actions
that should be present in the usage.
  (HasUsageActions): Indicate if at least one action is present in the usage.

* OmitFromUsageDocs: make it implement the IArgumentOrActionMetadata do make
the annotation usable for actions as well.

* UsageTemplates.cs: Make use of the news methods to only print actions if at
least one action is present.

*TableExpression.cs: filtrate the CommandLineArgument and CommentLineAction
instances that should be omitted from the evaluated elements.